add ta4j

+++ quant-service/src/main/java/com/optimus/mysql/entity/TradeCalendar.java
@@ -24,7 +24,7 @@ public class TradeCalendar extends BaseEntity {
 
     private String week;
 
-    private Integer isTrade;
+    private Short isTrade;
 
     private String holiday;
